pyproject.toml
[project]
name = "demogoogle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.11.0.post1",
    "google-cloud-aiplatform>=1.104.0",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.17",
    "langchain-community>=0.3.27",
    "langchain-google-vertexai>=2.0.27",
    "langchain-groq>=0.3.6",
    "langgraph>=0.5.3",
    "markdown>=3.8.2",
    "markdown2>=2.5.3",
    "openai>=1.97.0",
    "pillow>=11.3.0",
    "sentence-transformers>=5.0.0",
    "tiktoken>=0.9.0",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.53.3",
    "wikipedia>=1.4.0",
]
